+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Nag Mundari numeral system
 - `available_systems` function
+- `_validate_convert` function
 ### Changed
 - `README.md` updated
 - Test system modified